requirements.txt
pandas
python-dotenv
google-genai
rank_bm25
pymongo[srv]
pytest
google-cloud-firestore
rich
